+++ config.yml
@@ -2,6 +2,60 @@
 base_output_dir: "outputs"
 data_settings: "cfg_data.yml"
 
+data:
+  dataset_url: "https://raw.githubusercontent.com/Har-Lab/HumanActivityData/refs/heads/main/data/labeled_activity_data/"
+  dataset_name: "har-labs-activity-data"
+  raw_dir: "har_data"
+  processed_dir: "processed_data"
+  # "min_sample","subject", "None"
+  balance_setting: "subject"
+  window_center_percentage: 0.87
+  train_on_sensors: &sns
+    - ankle
+    - wrist
+    - waist
+  test_on_sensors:
+    # - ankle
+    - wrist
+    # - waist
+  classes: &activity_list
+    - jog_treadmill
+    - walk_treadmill
+    - upstairs
+    - downstairs
+    # - walk_mixed
+    # - walk_sidewalk
+  test_size: 0.25
+  val_size: 0.25
+  window_size: 250
+  stride: 5
+  ft_col:
+    - "x"
+    - "y"
+    - "z"
+    # - "vm"
+    # - "time"
+
+partition:
+  cross_set_configs: &cfg
+    activities_required: *activity_list
+    sensors_required: *sns
+    min_samples_per_activity: 1000
+    # age_range: [20, 30]
+    # injury_free: true
+
+  complete_data:
+    <<: *cfg
+
+  no_ankle_sensor:
+    <<: *cfg
+    sensors_required: ['wrist', 'waist']
+    sensors_excluded: ['ankle']
+
+  no_jogging_activity:
+    <<: *cfg
+    has_jogging: false
+
 # which models to test in this trial 
 # commenting out models will not test that model
 models_tested:
